+++ R/format_value.R
@@ -5,6 +5,7 @@
 #' @param protect_integers Should integers be kept as integers (i.e., without decimals)?
 #' @param missing Value by which \code{NA} values are replaced. By default, an empty string (i.e. \code{""}) is returned for \code{NA}.
 #' @param width Minimum width of the returned string. If not \code{NULL} and \code{width} is larger than the string's length, leading whitespaces are added to the string.
+#' @param as_percent Logical, if \code{TRUE}, value is formatted as percentage value.
 #' @param ... Arguments passed to or from other methods.
 #'
 #'
@@ -15,6 +16,8 @@
 #' format_value(1.2)
 #' format_value(1.2012313)
 #' format_value(c(0.0045, 234, -23))
+#' format_value(c(0.0045, .12, .34))
+#' format_value(c(0.0045, .12, .34), as_percent = TRUE)
 #'
 #' format_value(as.factor(c("A", "B", "A")))
 #' format_value(iris$Species)
@@ -24,31 +27,34 @@
 #'
 #' format_value(iris)
 #' @export
-format_value <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, ...) {
+format_value <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
   UseMethod("format_value")
 }
 
 
 #' @export
-format_value.data.frame <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, ...) {
-  as.data.frame(sapply(x, format_value, digits = digits, protect_integers = protect_integers, missing = missing, width = width, simplify = FALSE))
+format_value.data.frame <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
+  as.data.frame(sapply(x, format_value, digits = digits, protect_integers = protect_integers, missing = missing, width = width, as_percent = as_percent, simplify = FALSE))
 }
 
 
 #' @export
-format_value.numeric <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, ...) {
+format_value.numeric <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
   if (protect_integers) {
-    out <- .format_value_unless_integer(x, digits = digits, .missing = missing, .width = width, ...)
+    out <- .format_value_unless_integer(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, ...)
   } else {
-    out <- .format_value(x, digits = digits, .missing = missing, .width = width, ...)
+    out <- .format_value(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, ...)
   }
+
   # Deal with negative zeros
-  whitespace <- ifelse(is.null(width), "", " ")
-  out[out == "-0"] <- paste0(whitespace, "0")
-  out[out == "-0.0"] <- paste0(whitespace, "0.0")
-  out[out == "-0.00"] <- paste0(whitespace, "0.00")
-  out[out == "-0.000"] <- paste0(whitespace, "0.000")
-  out[out == "-0.0000"] <- paste0(whitespace, "0.0000")
+  if (!is.factor(x)) {
+    whitespace <- ifelse(is.null(width), "", " ")
+    out[out == "-0"] <- paste0(whitespace, "0")
+    out[out == "-0.0"] <- paste0(whitespace, "0.0")
+    out[out == "-0.00"] <- paste0(whitespace, "0.00")
+    out[out == "-0.000"] <- paste0(whitespace, "0.000")
+    out[out == "-0.0000"] <- paste0(whitespace, "0.0000")
+  }
   out
 }
 
@@ -68,9 +74,9 @@ format_value.logical <- format_value.numeric
 
 
 #' @importFrom stats na.omit
-.format_value_unless_integer <- function(x, digits = 2, .missing = "", .width = NULL, ...) {
+.format_value_unless_integer <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, ...) {
   if (is.numeric(x) && !all(.is.int(stats::na.omit(x)))) {
-    .format_value(x, digits = digits, .missing = .missing, .width = .width)
+    .format_value(x, digits = digits, .missing = .missing, .width = .width, .as_percent = .as_percent)
   } else if (anyNA(x)) {
     .convert_missing(x, .missing)
   } else {
@@ -80,9 +86,13 @@ format_value.logical <- format_value.numeric
 
 
 
-.format_value <- function(x, digits = 2, .missing = "", .width = NULL, ...) {
+.format_value <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, ...) {
   if (is.numeric(x)) {
-    x <- ifelse(is.na(x), .missing, ifelse(x > 1e+5, sprintf("%.5e", x), sprintf("%.*f", digits, x)))
+    if (isTRUE(.as_percent)) {
+      x <- ifelse(is.na(x), .missing, ifelse(x > 1e+5, sprintf("%.5e", x), sprintf("%.*f%%", digits, 100 * x)))
+    } else {
+      x <- ifelse(is.na(x), .missing, ifelse(x > 1e+5, sprintf("%.5e", x), sprintf("%.*f", digits, x)))
+    }
   } else if (anyNA(x)) {
     x <- .convert_missing(x, .missing)
   }